rename resvote_serer to resvote server
